Updated redux store files structure - Created `/src/_store/` folder - Moved `store.ts` - Moved `articlesSlice.ts` - Moved `settingsSlice.ts` - Moved `sideNavSlice.ts`
