+++ src/_store/slices/settingsSlice.ts
